VGA text might work, main is sudenlly complaining like a lil baby with no tummy rubs lol

+++ drivers/VGAtext.c
@@ -3,6 +3,6 @@
 void WriteChar(char character, unsigned int x, unsigned int y, unsigned int forecolour, unsigned int backcolour)
 {
 	unsigned int attribute = (backcolour << 4) | (forecolour & 0x0f);
-	volatile uint32_t* where = (volatile uint16_t*)VIDEO_MEMORY + (y * WIDTH + x);
+	volatile uint16_t* where = (volatile uint16_t*)VIDEO_MEMORY + (y * WIDTH + x);
 	*where = character | (attribute << 8);
 }
